# elm-frontend/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "吃饱了吗点餐平台",
    "description": "Tianjin University Software Engineering Project",
    "version": "1.0.0"
  },
...
  "paths": {
    "/api/businesses/{id}": {
      "get": {
        "summary": "getBusiness",
        "deprecated": false,
        "description": "",
        "operationId": "getBusiness",
        "tags": ["管理店铺"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultBusiness"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "updateBusiness",
        "deprecated": false,
        "description": "",
        "operationId": "updateBusiness",
        "tags": ["管理店铺"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Business"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultBusiness"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "deleteBusiness",
        "deprecated": false,
        "description": "",
        "operationId": "deleteBusiness",
        "tags": ["管理店铺"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    },
                    "data": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "patch": {
        "summary": "patchBusiness",
        "deprecated": false,
        "description": "",
        "operationId": "patchBusiness",
        "tags": ["管理店铺"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Business"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultBusiness"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/businesses": {
      "get": {
        "summary": "getBusinesses",
        "deprecated": false,
        "description": "",
        "operationId": "getBusinesses",
        "tags": ["管理店铺"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultListBusiness"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "addBusiness",
        "deprecated": false,
        "description": "",
        "operationId": "addBusiness",
        "tags": ["管理店铺"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Business"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultBusiness"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/orders/business/{id}": {
      "get": {
        "summary": "getOrdersByBusinessId",
        "deprecated": false,
        "description": "根据BusinessId获取该店铺的所有订单",
        "tags": ["管理订单"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "businessId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Order"
                      }
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/applications/business": {
      "post": {
        "summary": "提交店铺申请",
        "deprecated": false,
        "description": "商家申请店铺",
        "tags": ["管理店铺申请"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "createTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "updateTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "creator": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "updater": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "deleted": {
                    "type": "boolean"
                  },
                  "business": {
                    "$ref": "#/components/schemas/Business"
                  },
                  "handler": {
                    "$ref": "#/components/schemas/User"
                  },
                  "applicationExplain": {
                    "type": "string"
                  },
                  "applicationState": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "required": ["business", "id"]
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "format": "int64"
                        },
                        "createTime": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updateTime": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "creator": {
                          "type": "integer",
                          "format": "int64"
                        },
                        "updater": {
                          "type": "integer",
                          "format": "int64"
                        },
                        "deleted": {
                          "type": "boolean"
                        },
                        "business": {
                          "$ref": "#/components/schemas/Business"
                        },
                        "handler": {
                          "$ref": "#/components/schemas/User"
                        },
                        "applicationExplain": {
                          "type": "string"
                        },
                        "applicationState": {
                          "type": "integer",
                          "format": "int32"
                        }
                      },
                      "required": ["id"]
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "get": {
        "summary": "获取申请列表",
        "deprecated": false,
        "description": "获取所有店铺申请",
        "tags": ["管理店铺申请"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BusinessApplication"
                      }
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": ["data"]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/applications/business/{id}": {
      "get": {
        "summary": "获取申请详情",
        "deprecated": false,
        "description": "获取单个店铺申请",
        "tags": ["管理店铺申请"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/BusinessApplication"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": ["data"]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "patch": {
        "summary": "审批店铺申请",
        "deprecated": false,
        "description": "管理员审批店铺申请",
        "tags": ["管理店铺申请"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "createTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "updateTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "creator": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "updater": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "deleted": {
                    "type": "boolean"
                  },
                  "business": {
                    "$ref": "#/components/schemas/Business"
                  },
                  "handler": {
                    "$ref": "#/components/schemas/User"
                  },
                  "applicationExplain": {
                    "type": "string"
                  },
                  "applicationState": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "required": ["id", "applicationState"]
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/BusinessApplication"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": ["data"]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/reviews/business/{businessId}": {
      "get": {
        "summary": "获取店铺全部订单评价",
        "deprecated": false,
        "description": "获取店铺全部订单评价，若用户设置匿名，则order, customer, creator置null",
        "tags": ["管理评价"],
        "parameters": [
          {
            "name": "businessId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "null"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Review"
                      }
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
...
      "BusinessApplication": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "integer",
            "format": "int64"
          },
          "updater": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "business": {
            "$ref": "#/components/schemas/Business"
          },
          "handler": {
            "$ref": "#/components/schemas/User"
          },
          "applicationExplain": {
            "type": "string"
          },
          "applicationState": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": ["business", "id"]
      },
      "Authority": {
        "required": ["name"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "Business": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "integer",
            "format": "int64"
          },
          "updater": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "businessName": {
            "type": "string"
          },
          "businessOwner": {
            "$ref": "#/components/schemas/User"
          },
          "businessAddress": {
            "type": "string"
          },
          "businessExplain": {
            "type": "string"
          },
          "businessImg": {
            "type": "string"
          },
          "orderTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "startPrice": {
            "type": "number"
          },
          "deliveryPrice": {
            "type": "number"
          },
          "remarks": {
            "type": "string"
          }
        },
        "required": ["id"]
      },
      "Review": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "integer",
            "format": "int64"
          },
          "updater": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "customer": {
            "$ref": "#/components/schemas/User"
          },
          "business": {
            "$ref": "#/components/schemas/Business"
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          },
          "anonymous": {
            "type": "boolean"
          },
          "stars": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "string"
          }
        },
        "required": ["stars", "id"]
      },
      "User": {
        "required": ["username"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "integer",
            "format": "int64"
          },
          "updater": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "username": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "authorities": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Authority"
            }
          }
        }
      },
      "HttpResultBusiness": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Business"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "DeliveryAddress": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "integer",
            "format": "int64"
          },
          "updater": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "contactName": {
            "type": "string"
          },
          "contactSex": {
            "type": "integer",
            "format": "int32"
          },
          "contactTel": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "customer": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "integer",
            "format": "int64"
          },
          "updater": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "customer": {
            "$ref": "#/components/schemas/User"
          },
          "business": {
            "$ref": "#/components/schemas/Business"
          },
          "orderDate": {
            "type": "string",
            "format": "date-time"
          },
          "orderTotal": {
            "type": "number"
          },
          "deliveryAddress": {
            "$ref": "#/components/schemas/DeliveryAddress"
          },
          "orderState": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "HttpResultListBusiness": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Business"
            }
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [],
  "security": []
}